make TNFS FS
